+++ system/modules/core/classes/Ajax.php
@@ -298,7 +298,7 @@ class Ajax extends \Backend
 				// Convert the selected values
 				if ($varValue != '')
 				{
-					$varValue = trimsplit(',', $varValue);
+					$varValue = trimsplit("\t", $varValue);
 
 					// Automatically add resources to the DBAFS
 					if ($strKey == 'fileTree')