+++ system/modules/core/library/Contao/Automator.php
@@ -553,7 +553,8 @@ class Automator extends \System
 
 			foreach (scan(TL_ROOT . '/' . $strDir) as $strFile)
 			{
-				if (in_array($strFile, $arrFiles) || $strFile == '.htaccess')
+				// Ignore non PHP files and files which have been included before
+				if (substr($strFile, -4) != '.php' || in_array($strFile, $arrFiles))
 				{
 					continue;
 				}
@@ -612,7 +613,7 @@ class Automator extends \System
 
 				foreach (scan(TL_ROOT . '/' . $strDir) as $strFile)
 				{
-					if (in_array($strFile, $arrFiles) || $strFile == '.htaccess')
+					if ((substr($strFile, -4) != '.php' && substr($strFile, -4) != '.xlf') || in_array($strFile, $arrFiles))
 					{
 						continue;
 					}
@@ -691,7 +692,8 @@ class Automator extends \System
 
 			foreach (scan(TL_ROOT . '/' . $strDir) as $strFile)
 			{
-				if (in_array($strFile, $included) || $strFile == '.htaccess')
+				// Ignore non PHP files and files which have been included before
+				if (substr($strFile, -4) != '.php' || in_array($strFile, $included))
 				{
 					continue;
 				}