+++ system/modules/core/library/Contao/Database/Updater.php
@@ -584,7 +584,8 @@ class Updater extends \Controller
 
 			foreach (scan(TL_ROOT . '/' . $strDir) as $strFile)
 			{
-				if (in_array($strFile, $arrFiles) || $strFile == '.htaccess')
+				// Ignore non PHP files and files which have been included before
+				if (substr($strFile, -4) != '.php' || in_array($strFile, $arrFiles))
 				{
 					continue;
 				}