+++ system/modules/core/forms/FormPassword.php
@@ -152,9 +152,9 @@ class FormPassword extends \Widget
 		return sprintf('<label for="ctrl_%s_confirm" class="confirm%s">%s%s%s</label>',
 						$this->strId,
 						(strlen($this->strClass) ? ' ' . $this->strClass : ''),
-						($this->required ? '<span class="invisible">'.$GLOBALS['TL_LANG']['MSC']['mandatory'].'</span> ' : ''),
+						($this->mandatory ? '<span class="invisible">'.$GLOBALS['TL_LANG']['MSC']['mandatory'].'</span> ' : ''),
 						sprintf($GLOBALS['TL_LANG']['MSC']['confirmation'], $this->strLabel),
-						($this->required ? '<span class="mandatory">*</span>' : ''));
+						($this->mandatory ? '<span class="mandatory">*</span>' : ''));
 	}