+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.1.1 (2013-XX-XX)
 --------------------------
 
 ### Fixed
+Correctly bypass the cache if `bypassCache` is set (see #5872).
+
+### Fixed
 Preserve the CSS3PIE behavior file path when combining style sheets (see #5848).
 
 ### Fixed